LayoutTests/java/lc3/JSBoolean/boolean-005-n-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion boolean conversion

FAIL dt.setInteger( true ); dt.PUB_INT should throw an exception. Was 1.
FAIL dt.setInteger( true ); dt.getInteger() should throw an exception. Was 1.
PASS typeof dt.getInt() threw exception TypeError: Result of expression 'dt.getInt' [undefined] is not a function..
PASS successfullyParsed is true

TEST COMPLETE
 
